+++ packages/general/src/environment/index.ts
@@ -7,4 +7,5 @@
 export * from "./Environment.js";
 export * from "./Environmental.js";
 export * from "./RuntimeService.js";
+export * from "./ServiceBundle.js";
 export * from "./VariableService.js";