+++ packages/nodejs-shell/src/app.ts
@@ -5,10 +5,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Environment, LogDestination, LogFormat, Logger, LogLevel, singleton } from "#general";
+import { Environment, LogDestination, LogFormat, Logger, LogLevel } from "#general";
 import { createFileLogger } from "#nodejs";
-import { NodeJsBle } from "#nodejs-ble";
-import { Ble } from "#protocol";
+import "@matter/nodejs-ble";
 import yargs from "yargs/yargs";
 import { MatterNode } from "./MatterNode.js";
 import { Shell } from "./shell/Shell";
@@ -147,20 +146,13 @@ async function main() {
                 } else {
                     theShell = new Shell(theNode, nodeNum, PROMPT, process.stdin, process.stdout);
                 }
+
                 if (bleHciId !== undefined) {
-                    await theNode.Store.set("BleHciId", bleHciId);
+                    Environment.default.vars.set("ble.hci.id", bleHciId);
                 }
 
                 if (ble) {
-                    const hciId = await theNode.Store.get<number>("BleHciId", 0);
-                    // Initialize Ble
-                    Ble.get = singleton(
-                        () =>
-                            new NodeJsBle({
-                                environment: Environment.default,
-                                hciId,
-                            }),
-                    );
+                    Environment.default.vars.set("ble.enable", true);
                 }
 
                 console.log(`Started Node #${nodeNum} (Type: ${nodeType}) ${ble ? "with" : "without"} BLE`);